logging: backend: swo: add Kconfig option for SWO protocol
The TPIU supports serializing the data stream both using an UART-like NRZ
protocol as well as using Manchester encoding. Using Manchester encoding
has the advantage that it enables receivers that support it to recover the
clock from the SWO signal itself. This is particularly useful in situations
where the clock rate changes dynamically or is unknown (for example when
debugging the clock tree setup or working with a device with a misbehaving
main oscillator).

Add a Kconfig choice to switch the protocol, keeping the current default of
using the NRZ encoding.

(cherry picked from commit 22d6bfa0bdf20f3e44568695ebac2d4a4cf16403)

Original-Signed-off-by: Florian Larysch <larysch@fixme.gmbh>
GitOrigin-RevId: 22d6bfa0bdf20f3e44568695ebac2d4a4cf16403
Change-Id: I9e074431f4aa83ed5921d1d327fbb1bc10b83694
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5748678
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>