soc/mtk_adsp: Remove "msg" API
This is a feature of the 8195 DSP only, which is used only vestigially
by SOF to store data that nothing reads.  The Linux kernel on the
other side uses a shared driver for all 81xx devices, which does not
expose the feature.  It seems to work, but it's not worth maintaining
a driver in tree for legacy hardware that will never use it.

(cherry picked from commit cdd4cbcc676cac8ca3af6ba89528ba3ad779792b)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: cdd4cbcc676cac8ca3af6ba89528ba3ad779792b
Cr-Build-Id: 8729941316818268033
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8729941316818268033
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3d6426ff9032eb628dbac8d184f57641fd4dfe7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6053708
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>