drivers: sensor: ntc_thermistor: adds support for VDD based ADC reference
The NTC thermistor implementation assumes a constant pull-up voltage
and that the ADC channel is measured against a reference voltage so that
the absolute voltage across the NTC can be calculated. Based on the
relationship of `pullup-uv` and this absolute NTC voltage, the resistance
of the NTC is calculated.

There are applications where the "pullup-uv" is not constant, but VDD.
Most ADCs support relative measurements against VDD. If `pullup-uv` is not
defined, the implementation assumes now that the ADC channel is configured
to use VDD as a reference and therefore no millivolt conversion is
required.

(cherry picked from commit 8786436909b3d2ff7dd807f1504b44c1cbbecd18)

Original-Signed-off-by: Stefan Schwendeler <Stefan.Schwendeler@husqvarnagroup.com>
GitOrigin-RevId: 8786436909b3d2ff7dd807f1504b44c1cbbecd18
Cr-Build-Id: 8732749822481685601
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732749822481685601
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5983c84f54a81bc290fa1158a64310f8ee96f93a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5975815
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>