llext: set proper LLEXT_STORAGE default for Xtensa
This patch sets the default value for LLEXT_STORAGE_WRITABLE to 'y' on
the Xtensa architecture. This is necessary because it does not currently
support the read-only mode for the LLEXT storage.

Make sure the default reflects this instead of asking the user to
manually set it.

(cherry picked from commit 6c5983492ba94631fef1d608c5e65b52402a2421)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 6c5983492ba94631fef1d608c5e65b52402a2421
Change-Id: I6e0ae8909cc293360787cb2805c0a6389dec7802
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5658693
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>