llext: fix llext_find_section(), remove llext_section_by_name()
The function llext_section_by_name() is used only in one place, and it
expects the caller to have the section headers cache available. This
cache is freed after the ELF file is loaded, so the function is not
usable in the context where it is called.

Remove the function and replace the call with a direct search in the
ELF file section headers array, as was done before 08eb314c35d.

(cherry picked from commit d4ea1da10e6c69911218263115a307f0336aa666)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: d4ea1da10e6c69911218263115a307f0336aa666
Change-Id: Id14ebf2b343af1e29f6af1e4c6073cbe74760102
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5658699
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>