Bluetooth: test: privacy/legacy: Scan continuously
Some controllers, like the SoftDevice Controller can miss the first few
advertisements after the host tells it to start scanning. The DUT would
assumes the first RPA advertisements report it gets is the first that
was sent, but it was actually not the first. This would skew the
tester's judgement about the timing of RPA rotations.

To remedy this, the tester will scan continuously and switch between
expecting the identity address and the RPA without stopping the scanner.

The tester will tell the DUT to switch to RPA after it has received the
identity address advertisement. This ensures that the tester will not
miss the first RPA advertisement.

(cherry picked from commit 8cd98b139d252736da63c8a016b23dff3c163935)

Original-Signed-off-by: Aleksander Wasaznik <aleksander.wasaznik@nordicsemi.no>
GitOrigin-RevId: 8cd98b139d252736da63c8a016b23dff3c163935
Change-Id: Ied376c6e31ca3786c7d9650c33a1656441938bc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5667228
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>