drivers: dp: swdp_bitbang: power optimization
This patch changes GPIO initialization
to be in PORT_OFF state by default.
Also, if no transceiver is attached to the signals,
the GPIOs are configured to be disconnected to preserve power.

I tested this on a prototype board where we are going to have
a debugger in a low-power context.

(cherry picked from commit b0936ae35331c3218ed4d51477d95b06afe53ecd)

Original-Signed-off-by: Maximilian Deubel <maximilian.deubel@nordicsemi.no>
GitOrigin-RevId: b0936ae35331c3218ed4d51477d95b06afe53ecd
Cr-Build-Id: 8734199375406094193
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8734199375406094193
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iba6ff927aea7e73db9a216861037e20eb0eafb2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5927190
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>