drivers: nxp_enet: Fix CID 363724
This addresses the coverity issue ID 363724.
This issue seems to have been a false positive, but the root complaint
from coverity was that within the callstack of ENET_Up, some functions
use the buffer config pointer as an array, whereas in the zephyr driver
it is declared as a singleton. Address it by redeclaring this as an
array of size 1 instead of a singleton.

(cherry picked from commit 3f59cf9439938c6f5054e86f7f396074ff5685cc)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 3f59cf9439938c6f5054e86f7f396074ff5685cc
Change-Id: I146f91aa9fc4a53a7b862a44b55dfbfccfe6382d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5663138
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Ting Shen <phoenixshen@chromium.org>