ci: set_assignee: pick next area when submitter = assignee
Address the cases where submitter is also the maintainer of the code
changed and other areas are being changed. In this case, assign to the
next area maintainers instead of assigning to submitter.

area A but also makes changes to other areas B and C. Right now
maintainers of B and C are added as reviewers.

This change will assign to the next area after A, i.e. B in cases where the
submitter is also the maintainer of area A.

(cherry picked from commit 913426e585e9dc72b9bc3eb04f44be192ccbfd6a)

Original-Example: maintainer of component A introduced significant changes to
Original-Signed-off-by: Anas Nashif <anas.nashif@intel.com>
GitOrigin-RevId: 913426e585e9dc72b9bc3eb04f44be192ccbfd6a
Change-Id: I3fcd5b9f1b62f658e93b55eb2aba21750459f9ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5694419
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>