usb: fix wrong capacity report for USB mass storage device
Usb "write" method never copies the input data buffer, it uses the
buffer pointer directly. The fact that it was on the stack made
the buffer ephemeral, and could disappear before the USB transfer
actually occurred. The data transmitted was then random. This commit
converts the transfer buffer to a static buffer, so that it always
exists for the duration of the USB transfer.

Known side-effect: if "read capacity" command is read twice
simultaneously, the same buffer will be used twice. As the capacity of a
USB drive should not change between 2 calls, this side effect can be
ignored.

(cherry picked from commit 32481b64185205222a58acdc6e7272ef2c06869d)

Original-Signed-off-by: Alexi Demers <alexi.demers@axceta.com>
GitOrigin-RevId: 32481b64185205222a58acdc6e7272ef2c06869d
Cr-Build-Id: 8738457431847540977
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8738457431847540977
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie58861caa15ab48a1a64020c97d019e634b419f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5817303
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>