drivers: sensor: ds18b20: indicate error for missing response
When multiple devices are on the bus, and no slave matches
the sent rom, the master receives only ones.
This should be handled explicitly and the error reporting
should not rely on the crc mismatch of the scratchpad.

eeprom (copy scratchpad), otherwise the resolution will not match
the expectations in case a node looses power and recovers.
As in that case the configuration will not be applied again.
Also the sensor may still report the power on reset value
of +85°C just after the sensor powered up.

(cherry picked from commit b5d3dea8943e48580acb922932146a978974c145)

Original-Note: Make sure that the sensor configuration is loaded to the
Original-Signed-off-by: Thomas Stranger <thomas.stranger@outlook.com>
GitOrigin-RevId: b5d3dea8943e48580acb922932146a978974c145
Change-Id: I60d8ffef438556a2b042abbee3f17ad5b2ff05bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5691903
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>