drivers: eeprom: emulator: correct improper exponentiation
The top-level driver comment states that a uint16_t is used to store
changes if the EEPROM size is smaller than 64kB, or uint32_t otherwise.
However, the 64kB threshold is represented as 2^16 in the actual code,
which is a bitwise XOR instead of an exponent. This causes a uint16_t
to be used only if the size is less than or equal to 18 bytes. Correct
this by using KB(64) instead of 2^16.

(cherry picked from commit 34e4edc28a3806885821d863cc3ebc9c60f5b817)

Original-Signed-off-by: Jared Kangas <kangas.jd@gmail.com>
GitOrigin-RevId: 34e4edc28a3806885821d863cc3ebc9c60f5b817
Change-Id: I2f31a71224fda374e6d65d08d23112901f7b5609
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5680110
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>