net: sockets: Remove async service support
As found in PR #75525, we should not modify the polled fd array
in multiple places. Because of this fix, the async version of
the socket service could start to trigger while it is being handled
by the async handler. This basically means that the async version
cannot work as intended so remove its support.

(cherry picked from commit 794d7cf3c5554991cf12788049feb75537080817)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: 794d7cf3c5554991cf12788049feb75537080817
Change-Id: I049890a7827d3a03f245abf171877f4ab0304e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5691930
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>