logging: Fix runtime filtering when frontend and userspace is used
When userspace is used and frontend was used for logging then runtime
filtering was failing because in user context filtering data was
accessed and filtering data is in the kernel space. Fixing that and
adding runtime filtering to the pre frontend function which is
already executed in the kernel space and filter data can be
accessed.

(cherry picked from commit 347eeababf50f2eb98be1fbb948d335834aeec33)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 347eeababf50f2eb98be1fbb948d335834aeec33
Change-Id: I58d4ad204debf9afbdd97a8f72c5eb3445093fef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5689408
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>