net: lib: http_server: Respect Connection close header
In case no "Connection: close" header is present in the request, the
server should keep the connection open for the client. Hence, after
serving a request, we need to check if the header was present (the
parser sets a flag for it), and only close the connection immediately,
if the client requested it. In case the client remains silent, the
connection will be closed anyway after the inactivity timer kicks in.

(cherry picked from commit 5251533f975cc88e79a61bb3c5259d814e3838c8)

Original-Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
GitOrigin-RevId: 5251533f975cc88e79a61bb3c5259d814e3838c8
Change-Id: I1d7ed3558d53bafe22504b2575813ac755224a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5702147
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>