twister/pytest: Add unlaunched_dut fixture
Add a new fixture, `unlaunched_dut`, which is basically the `dut`
one, but without launching the device - thus also not building the
application. It is useful for tests who need a finer control of
the building of the application, before the dut can be launched
to run it.

It will be used on a future patch, which will use it to enable LLEXT EDK
tests.

(cherry picked from commit 55613d035bcc2663388b72f491d89494e15ec072)

Original-Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>
GitOrigin-RevId: 55613d035bcc2663388b72f491d89494e15ec072
Cr-Build-Id: 8739453998936228337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8739453998936228337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie9a49e17aa47f43ee44eca0c0a86a586efa32825
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/5790537
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>